requirements.txt
django==4.2.7
django-extensions
ipython
Jinja2
django-debug-toolbar
django-admin-rangefilter
django-import-export
django-storages
django-filter
djangorestframework
djangorestframework-xml
djangorestframework-yaml
djangorestframework-simplejwt
drf-yasg
boto3
Celery
requests
model-bakery

#flake8
flake8
flake8-builtins
flake8-print
flake8-breakpoint

#pytest
pytest
pytest-django
pytest-cov
pytest-mock
